Remove incomplete API infrastructure to connect to MiniZinc
With the current developments within MiniZinc it seems unlikely that
the framework as currently implemented will still be used to directly
connect to the current version of libminizinc. As such, it makes sense
to remove these features for now. This will simplify the Python library
and likely allow us to give more sensible errors.
