--- src/minizinc/CLI/__init__.py
@@ -1,8 +1,0 @@
-#  This Source Code Form is subject to the terms of the Mozilla Public
-#  License, v. 2.0. If a copy of the MPL was not distributed with this
-#  file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-from .driver import CLIDriver
-from .instance import CLIInstance
-
-__all__ = ["CLIDriver", "CLIInstance"]
